test(e2e): split by domain and add shared setup helpers
- Replace single client.test.ts with domain-specific files (oracle, config,
  cdp, flash, psm, saving, savings-rewards, supply-pools, vaults, etc.)
- Add shared setup helpers in test/e2e/helpers/setup.ts

Made-with: Cursor
